Refactor de code : Modification de l'arborescance du projet
- les views sont dans le dossier views
- les forms sont dans le dossier forms
- les models sont dans le dossier models
