+++ formation/views/__init__.py
@@ -1,0 +1,15 @@
+from .calendar_view import *
+from .cancel_registration_session import *
+from .detail_session_view import *
+from .detail_formation_view import *
+from .formation_list_view import *
+from .formation_delete_view import *
+from .formation_list_for_formateur_view import *
+from .new_session_form_view import *
+from .new_formation_form_view import *
+from .new_registration_form_view import *
+from .registration_session import *
+from .registration_list_for_students_view import *
+from .session_delete_view import *
+from .update_session_view import *
+from .update_formation_view import *
